fixed issue 71
